requirements.txt
streamlit ==1.15.1
plotly==5.10.0
pandas==1.4.3
numpy==1.23.1
folium0.13.0
matplotlib=3.5.3
matplotlib-inline==0.1.6
Pillow==9.2.0